New ragdoll (#359)
* New Blip features (#292)

* Update GTAblip.cpp

⚠VERY EXPERIMENTAL⚠

I have tried to add some new features to blips. Please recognize that I have never coded in my life and this is all black magic to me.

It would mean a lot to me if this could be tested!

Added;
Toggle for cones (like NPC vision, likely won't function but it's there for the aesthetic).

Toggle to make blips not selectable on the map.

Set blip priority.

Toggle to allow blip rotation.

Toggle (hopefully just a toggle) to have a blip to mark an area.

Changes;
Moved Enemy, Friend, and VIP to the top

* Update GTAblip.h

⚠VERY EXPERIMENTAL⚠
Additions made to this code were added by someone with little to no coding experience.

See commit to GTAblip.cpp for a full list of changes.

* Update GTAblip.h

* Added one last component.

Added the function for blips to rotate in sync with the attached entity (in theory).

* Update STSTasks.cpp

Adding handles.

* Update STSTasks.h

* Update Submenus_TaskSequence.cpp

* Update GTAblip.cpp

* Update STSTasks.cpp

* Update STSTasks.cpp

* Update STSTasks.cpp

* Update STSTasks.cpp

* Update STSTasks.cpp

* Update STSTasks.cpp

* Edit Ragdoll input values

Ped Ragdoll switched to a better type that actually can't fall.

* Blips Update & Landing Gear (#314)

* Update GTAblip.cpp

⚠VERY EXPERIMENTAL⚠

I have tried to add some new features to blips. Please recognize that I have never coded in my life and this is all black magic to me.

It would mean a lot to me if this could be tested!

Added;
Toggle for cones (like NPC vision, likely won't function but it's there for the aesthetic).

Toggle to make blips not selectable on the map.

Set blip priority.

Toggle to allow blip rotation.

Toggle (hopefully just a toggle) to have a blip to mark an area.

Changes;
Moved Enemy, Friend, and VIP to the top

* Update GTAblip.h

⚠VERY EXPERIMENTAL⚠
Additions made to this code were added by someone with little to no coding experience.

See commit to GTAblip.cpp for a full list of changes.

* Update GTAblip.h

* Added one last component.

Added the function for blips to rotate in sync with the attached entity (in theory).

* Update STSTasks.cpp

Adding handles.

* Update STSTasks.h

* Update Submenus_TaskSequence.cpp

* Update STSTasks.cpp

Added more options to cones.

* Update STSTasks.cpp

I forgot to add to the XML Node...

* Update STSTasks.h

Added value defaults.

* Update GTAblip.h

* Update GTAblip.h

* Update GTAblip.h

* Update GTAblip.cpp

* Update STSTasks.cpp

* Update STSTasks.h

* Update STSTasks.h

Fixing issues with Blip "Display" Native being buggy.

* Update STSTasks.cpp

Fixing issues with SET_BLIP_DISPLAY.

* Update GTAblip.cpp

* Update GTAblip.h

* Update Routine.cpp

I decided to commit to this version of no ragdoll I made.

It may be fruitful to add an option to change "no ragdoll" styles, but for right now I don't think it is worth the effort, so this improved version will do for now.

* Update GTAentity.cpp

Added a landing gear toggle for aircraft.

One day I want to include landing gear and boat anchors as tasks, but that is not today.

At some point I also want to try to move boat anchor toggle to here too.

* Update GTAentity.h

Added landing gear .h necessities.

---------

Co-authored-by: ThatRealSpicyDragon <158034338+ThatRealSpicyDragon@users.noreply.github.com>